changed template type of Shape, Strides, NDIndex & SliceRanges to uint8_t

+++ laruen/ndarray/ndarray_types.h
@@ -11,14 +11,14 @@ using namespace laruen::utils;
 typedef float float32_t;
 typedef double float64_t;
 
-template <uint64_t N> using Shape = Array<uint32_t, N>;
-template <uint64_t N> using Strides = Array<uint64_t, N>;
-template <uint64_t N> using NDIndex = Array<uint32_t, N>;
-template <uint64_t N> using SliceRanges = Array<Range<uint32_t>, N>;
+template <uint8_t N> using Shape = Array<uint32_t, N>;
+template <uint8_t N> using Strides = Array<uint64_t, N>;
+template <uint8_t N> using NDIndex = Array<uint32_t, N>;
+template <uint8_t N> using SliceRanges = Array<Range<uint32_t>, N>;
 
-template <uint64_t N> std::ostream& operator<<(std::ostream &strm, const Shape<N> &shape);
-template <uint64_t N> std::ostream& operator<<(std::ostream &strm, const Strides<N> &strides);
-template <uint64_t N> std::ostream& operator<<(std::ostream &strm, const SliceRanges<N> &slice_ranges);
+template <uint8_t N> std::ostream& operator<<(std::ostream &strm, const Shape<N> &shape);
+template <uint8_t N> std::ostream& operator<<(std::ostream &strm, const Strides<N> &strides);
+template <uint8_t N> std::ostream& operator<<(std::ostream &strm, const SliceRanges<N> &slice_ranges);
 
 #include "laruen/ndarray/ndarray_types.tpp"
 #endif
